+++ jest.config.js
@@ -1,13 +1,19 @@
 module.exports = {
   moduleNameMapper: {
     css$: 'identity-obj-proxy',
-    svg$: 'identity-obj-proxy',
   },
-  // Key is deprecated in future version in favor of setupFilesAfterEnv
-  setupTestFrameworkScriptFile: '<rootDir>/setup-test-framework.js',
+  setupFilesAfterEnv: [
+    '@testing-library/jest-dom/extend-expect.js',
+    '<rootDir>/jest.config.setup.files.after.env.js',
+  ],
   testPathIgnorePatterns: [
     'node_modules',
     'styleguide/cypress/integration',
   ],
+  transform: {
+    js$: 'babel-jest',
+    jsx$: 'babel-jest',
+    svg$: '<rootDir>/jest.config.transform.svg.js',
+  },
   verbose: true,
 };