+++ src/components/custom-field-input-number/custom-field-input-number.test.jsx
@@ -1,74 +1,104 @@
 import React from 'react';
-import { render, fireEvent } from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
 import renderer from 'react-test-renderer';
-import CustomFieldInputText from '../custom-field-input-text/custom-field-input-text.jsx';
 import CustomFieldInputNumber from './custom-field-input-number.jsx';
 
 describe('CustomFieldInputNumber', () => {
+  function TestComponent(props = {}) {
+    return <CustomFieldInputNumber id="test-input" label="Test label" {...props} />;
+  }
+
   it('has defaults', () => {
-    const tree = renderer.create((
-      <CustomFieldInputNumber />
-    )).toJSON();
+    const tree = renderer.create(<TestComponent />).toJSON();
     expect(tree).toMatchSnapshot();
   });
 
-  describe('prop-forward API', () => {
-    it('forwards all props accepted by CustomFieldInputText on Object keys', () => {
-      const numberProps = Object.keys(CustomFieldInputNumber.propTypes);
-      Object.keys(CustomFieldInputText.propTypes).forEach((key) => {
-        expect(numberProps).toContain(key);
-      });
-    });
-
-    it('renders all forwarded props except event handlers', () => {
-      const tree = renderer.create((
-        <CustomFieldInputNumber
-          className={'test'}
-          disabled={false}
-          error={true}
-          helpText={'help'}
-          id={'id'}
-          name={'name'}
-          placeholder={'placeholder'}
-          required={true}
-          type={'text'}
-          value={'value'}
-        />
-      )).toJSON();
-      const stringTree = JSON.stringify(tree);
-
-      expect(tree.props.className).toContain('test error');
-      expect(tree.props.className).not.toContain('disabled');
-
-      expect(stringTree).toContain('help');
-      expect(stringTree).toContain('id');
-      expect(stringTree).toContain('name');
-      expect(stringTree).toContain('placeholder');
-      expect(stringTree).toContain('Required');
-      expect(stringTree).toContain('type');
-      expect(stringTree).toContain('value');
+  describe('className API', () => {
+    it('prioritizes className prop', () => {
+      const { container } = render(<TestComponent className="prioritize-me" />);
+      expect(container.firstChild).toHaveClass('prioritize-me');
     });
   });
 
-  describe('input validation', () => {
-    it('does not trigger error state for valid value', () => {
-      const { getByRole, getByTestId } = render(
-        <CustomFieldInputNumber />,
-      );
-
-      fireEvent.change(getByRole('textbox'), { target: { value: '1234' } });
-
-      expect(getByTestId('custom-field-input')).not.toHaveClass('error');
+  describe('disabled API', () => {
+    it('can be disabled', () => {
+      const { container } = render(<TestComponent disabled />);
+      expect(container.firstChild).toHaveClass('disabled');
+      expect(screen.getByLabelText('Test label')).toBeDisabled();
     });
 
-    it('has error state for invalid value', () => {
-      const { getByRole, getByTestId } = render(
-        <CustomFieldInputNumber />,
-      );
+    it('can be enabled', () => {
+      const { container } = render(<TestComponent />);
+      expect(container.firstChild).not.toHaveClass('disabled');
+      expect(screen.getByLabelText('Test label')).not.toBeDisabled();
+    });
+  });
 
-      fireEvent.change(getByRole('textbox'), { target: { value: '--0.1.2' } });
+  describe('id API', () => {
+    it('sets the id attribute', () => {
+      render(<TestComponent id="test-id" />);
+      expect(screen.getByLabelText('Test label')).toHaveAttribute('id', 'test-id');
+    });
+  });
 
-      expect(getByTestId('custom-field-input')).toHaveClass('error');
+  describe('label API', () => {
+    it('sets the label', () => {
+      render(<TestComponent label="Another label" />);
+      expect(screen.getByLabelText('Another label')).toBeDefined();
+    });
+  });
+
+  describe('name API', () => {
+    it('sets the name attribute', () => {
+      render(<TestComponent name="test-name" />);
+      expect(screen.getByLabelText('Test label')).toHaveAttribute('name', 'test-name');
+    });
+  });
+
+  describe('placeholder API', () => {
+    it('sets the placeholder attribute', () => {
+      const placeholder = 'This is placeholder input';
+      render(<TestComponent placeholder={placeholder} />);
+      expect(screen.getByLabelText('Test label')).toHaveAttribute('placeholder', placeholder);
+    });
+  });
+
+  describe('required API', () => {
+    it('sets the required attribute', () => {
+      render(<TestComponent required />);
+      expect(screen.getByLabelText('Test label')).toBeRequired();
+    });
+
+    it('unsets the required attribute', () => {
+      render(<TestComponent />);
+      expect(screen.getByLabelText('Test label')).not.toBeRequired();
+    });
+  });
+
+  describe('value API', () => {
+    it('is valid on a postive integer', () => {
+      render(<TestComponent value={1} />);
+      expect(screen.getByTestId('custom-field-input')).not.toHaveClass('error');
+    });
+
+    it('is valid on zero', () => {
+      render(<TestComponent value={0} />);
+      expect(screen.getByTestId('custom-field-input')).not.toHaveClass('error');
+    });
+
+    it('is valid on a negative integer', () => {
+      render(<TestComponent value={-1} />);
+      expect(screen.getByTestId('custom-field-input')).not.toHaveClass('error');
+    });
+
+    it('is valid on a decimal integer', () => {
+      render(<TestComponent value={1.00} />);
+      expect(screen.getByTestId('custom-field-input')).not.toHaveClass('error');
+    });
+
+    it('is invalid on a decimal number', () => {
+      render(<TestComponent value={1.01} />);
+      expect(screen.getByTestId('custom-field-input')).toHaveClass('error');
     });
   });
 });