+++ src/components/custom-field-input-text/custom-field-input-text.test.jsx
@@ -1,192 +1,158 @@
-import React from 'react';
-import { render, fireEvent } from '@testing-library/react';
+import React, { createRef } from 'react';
 import renderer from 'react-test-renderer';
+import { fireEvent, render, screen } from '@testing-library/react';
 import CustomFieldInputText from './custom-field-input-text.jsx';
 
 describe('CustomFieldInputText', () => {
+  function TestComponent(props = {}) {
+    return <CustomFieldInputText id="test-input" label="Test label" {...props} />;
+  }
+
   it('has defaults', () => {
-    const tree = renderer.create((
-      <CustomFieldInputText />
-    )).toJSON();
-    expect(tree).toMatchSnapshot();
+    expect(renderer.create(<TestComponent />).toJSON()).toMatchSnapshot();
   });
 
   describe('className API', () => {
     it('prioritizes className prop', () => {
-      const tree = renderer.create((
-        <CustomFieldInputText className="prioritize-me" />
-      )).toJSON();
-      expect(tree.props.className).toEqual('prioritize-me');
+      const { container } = render(<TestComponent className="prioritize-me" />);
+      expect(container.firstChild).toHaveClass('prioritize-me');
     });
   });
 
   describe('disabled API', () => {
     it('can be disabled', () => {
-      const r = renderer.create((
-        <CustomFieldInputText disabled={true} />
-      ));
-      const tree = r.toJSON();
-      const root = r.root;
-
-      expect(tree.props.className).toContain('disabled');
-      expect(root.findByType('input').findByProps({ disabled: true }));
+      const { container } = render(<TestComponent disabled />);
+      expect(container.firstChild).toHaveClass('disabled');
+      expect(screen.getByLabelText('Test label')).toBeDisabled();
     });
 
     it('can be enabled', () => {
-      const r = renderer.create((
-        <CustomFieldInputText disabled={false} />
-      ));
-      const tree = r.toJSON();
-      const root = r.root;
-
-      expect(tree.props.className).not.toContain('disabled');
-      expect(root.findByType('input').findByProps({ disabled: false }));
+      const { container } = render(<TestComponent />);
+      expect(container.firstChild).not.toHaveClass('disabled');
+      expect(screen.getByLabelText('Test label')).not.toBeDisabled();
     });
   });
 
-  describe('error API', () => {
-    it('can have an error state', () => {
-      const r = renderer.create((
-        <CustomFieldInputText error={true} />
-      ));
-      const tree = r.toJSON();
-      const root = r.root;
+  describe('error API and helpText API', () => {
+    xit('can have an error state through a native validation', () => {
+      // I am not sure what is the best way to represent this in a test.
+      // However, at the moment, there are end-to-end tests in the Number component tests.
+      const { container } = render(<TestComponent error />);
+      expect(container.firstChild).toHaveClass('error');
+      expect(screen.getByLabelText('Test label')).toBeInvalid();
+      expect(screen.getByRole('img').firstChild).toHaveAttribute('xlink:href', '#icon-caution-fill.svg');
+    });
 
-      expect(tree.props.className).toContain('error');
-      expect(root.findByProps({ currentColor: 'caution' }));
+    it('can have an error state through a custom validation', () => {
+      const { container } = render(<TestComponent error helpText="Custom validation message" />);
+      expect(container.firstChild).toHaveClass('error');
+      expect(screen.getByLabelText('Test label')).toBeInvalid();
+      expect(screen.getByRole('img').firstChild).toHaveAttribute('xlink:href', '#icon-caution-fill.svg');
     });
 
     it('can have no error state', () => {
-      const r = renderer.create((
-        <CustomFieldInputText error={false} />
-      ));
-      const tree = r.toJSON();
-      const root = r.root;
-
-      expect(tree.props.className).not.toContain('error');
-      expect(root.findByProps({ className: 'input-container' }).children.length).toEqual(1);
+      const { container } = render(<TestComponent />);
+      expect(container.firstChild).not.toHaveClass('error');
+      expect(screen.getByLabelText('Test label')).toBeValid();
+      expect(container.querySelector('[role="img"]')).toBeFalsy();
     });
   });
 
-  describe('help text API', () => {
-    it('can have help text', () => {
-      const helpText = 'Oh wow big helpful yes!';
-      const r = renderer.create((
-        <CustomFieldInputText helpText={helpText} />
-      ));
-      const root = r.root;
-
-      expect(root.findByProps({ className: 'help' }).children).toContain(helpText);
+  describe('id API', () => {
+    it('sets the id attribute', () => {
+      render(<TestComponent id="test-id" />);
+      expect(screen.getByLabelText('Test label')).toHaveAttribute('id', 'test-id');
     });
+  });
 
-    it('can have empty help text', () => {
-      const r = renderer.create((
-        <CustomFieldInputText />
-      ));
-      const root = r.root;
+  describe('inputRef API', () => {
+    it('sets the ref on the input', () => {
+      const inputRef = createRef();
+      render(<TestComponent inputRef={inputRef} />);
+      expect(screen.getByLabelText('Test label')).toBe(inputRef.current);
+    });
+  });
 
-      expect(root.findByProps({ className: 'help' }).children).toEqual([]);
+  describe('label API', () => {
+    it('sets the label', () => {
+      render(<TestComponent label="Another label" />);
+      expect(screen.getByLabelText('Another label')).toBeDefined();
+    });
+  });
+
+  describe('max API', () => {
+    it('sets the max attribute', () => {
+      render(<TestComponent max={5} />);
+      expect(screen.getByLabelText('Test label')).toHaveAttribute('max', '5');
+    });
+  });
+
+  describe('min API', () => {
+    it('sets the min attribute', () => {
+      render(<TestComponent min={-5} />);
+      expect(screen.getByLabelText('Test label')).toHaveAttribute('min', '-5');
     });
   });
 
   describe('name API', () => {
     it('sets the name attribute', () => {
-      const tree = renderer.create((
-        <CustomFieldInputText name="test-name" />
-      )).root;
-      expect(tree.findByType('input').findByProps({ name: 'test-name' }));
+      render(<TestComponent name="test-name" />);
+      expect(screen.getByLabelText('Test label')).toHaveAttribute('name', 'test-name');
+    });
+  });
+
+  describe('onKeyUp', () => {
+    it('handles the key up event', () => {
+      const onKeyUpSpy = jest.fn();
+      render(<TestComponent onKeyUp={onKeyUpSpy} />);
+      fireEvent.keyUp(screen.getByLabelText('Test label'));
+      expect(onKeyUpSpy.mock.calls.length).toBe(1);
     });
   });
 
   describe('placeholder API', () => {
     it('can have placeholder for input', () => {
       const placeholder = 'This is placeholder input';
-      const r = renderer.create((
-        <CustomFieldInputText placeholder={placeholder} />
-      ));
-      const root = r.root;
-
-      expect(root.findByType('input').props.placeholder).toEqual(placeholder);
-    });
-
-    it('can have no placeholder for input', () => {
-      const r = renderer.create((
-        <CustomFieldInputText />
-      ));
-      const root = r.root;
-
-      expect(root.findByType('input').props.placeholder).toBeUndefined();
+      render(<TestComponent placeholder={placeholder} />);
+      expect(screen.getByLabelText('Test label')).toHaveAttribute('placeholder', placeholder);
     });
   });
 
   describe('required API', () => {
     it('can have a required indicator', () => {
-      const r = renderer.create((
-        <CustomFieldInputText required={true} />
-      ));
-      const root = r.root;
-
-      expect(root.findByProps({ className: 'optional' }));
+      render(<TestComponent required />);
+      expect(screen.getByLabelText('Test label')).toBeRequired();
     });
 
     it('can have no required indicator', () => {
-      const r = renderer.create((
-        <CustomFieldInputText />
-      ));
-      const root = r.root;
-
-      expect(root.findByProps({ className: 'input-container' }).children.length).toEqual(1);
+      render(<TestComponent />);
+      expect(screen.getByLabelText('Test label')).not.toBeRequired();
     });
   });
 
-  describe('id API', () => {
-    it('sets the id attribute', () => {
-      const tree = renderer.create((
-        <CustomFieldInputText id="test-id" />
-      )).root;
-      expect(tree.findByType('input').findByProps({ id: 'test-id' }));
+  describe('step API', () => {
+    it('sets the step attribute', () => {
+      render(<TestComponent step={2} />);
+      expect(screen.getByLabelText('Test label')).toHaveAttribute('step', '2');
     });
   });
 
   describe('type API', () => {
-    it('sets the type attribute', () => {
-      const tree = renderer.create((
-        <CustomFieldInputText type="button" />
-      )).root;
-      expect(tree.findByType('input').findByProps({ type: 'button' }));
+    it('can be set to `number`', () => {
+      render(<TestComponent type="number" />);
+      expect(screen.getByLabelText('Test label')).toHaveAttribute('type', 'number');
     });
-  });
 
-  describe('onChange API', () => {
-    it('sets and fires the onchange handler', () => {
-      const onChangeSpy = jest.fn();
-      const { getByRole } = render(
-        <CustomFieldInputText onChange={onChangeSpy} />,
-      );
-
-      fireEvent.change(getByRole('textbox'), { target: { value: 'this is text' } });
-
-      expect(onChangeSpy.mock.calls.length).toEqual(1);
-    });
-  });
-
-  describe('onClick API', () => {
-    it('sets and fires the onclick handler', () => {
-      const onClickSpy = jest.fn();
-      const tree = renderer.create((
-        <CustomFieldInputText onClick={onClickSpy} />
-      )).root;
-      tree.findByType('input').props.onClick();
-      expect(onClickSpy.mock.calls.length).toEqual(1);
+    it('can be set to `text`', () => {
+      render(<TestComponent type="text" />);
+      expect(screen.getByLabelText('Test label')).toHaveAttribute('type', 'text');
     });
   });
 
   describe('value API', () => {
     it('sets the value attribute', () => {
-      const tree = renderer.create((
-        <CustomFieldInputText value="test-value" />
-      )).root;
-      expect(tree.findByType('input').findByProps({ value: 'test-value' }));
+      render(<TestComponent value="test-value" />);
+      expect(screen.getByLabelText('Test label')).toHaveValue('test-value');
     });
   });
 });